+++ geoCosiCorr3D/utils/misc.py
@@ -43,4 +43,42 @@ def log_available_memory(component_name: str):
     total_memory_gb = memory_stats.total / (1024 ** 3)
     available_memory = memory_stats.available / (1024 ** 3)
     logging.info(f'{component_name}: _memory [Gb] : {available_memory:.2f} / {total_memory_gb:.3f}')
-    return  available_memory
+    return available_memory
+
+
+def compute_tile_fp(east_arr, north_arr, grid_epsg, tile_num):
+    import json
+    min_east = np.min(east_arr)
+    max_east = np.max(east_arr)
+    min_north = np.min(north_arr)
+    max_north = np.max(north_arr)
+
+    footprint_corners = [
+        [min_east, max_north],  # Top-left
+        [max_east, max_north],  # Top-right
+        [max_east, min_north],  # Bottom-right
+        [min_east, min_north],  # Bottom-left
+        [min_east, max_north]
+    ]
+    geojson_object = {
+        "type": "FeatureCollection",
+        "features": [
+            {
+                "type": "Feature",
+                "geometry": {
+                    "type": "Polygon",
+                    "coordinates": [footprint_corners]
+                },
+                "properties": {}
+            }
+        ],
+        "crs": {
+            "type": "name",
+            "properties": {
+                "name": f"urn:ogc:def:crs:EPSG::{grid_epsg}"
+            }
+        }
+    }
+    output_file_path = f'footprint_{tile_num}.geojson'
+    with open(output_file_path, 'w') as f:
+        json.dump(geojson_object, f)
